+++ app/(ui)/u2f/set/page.tsx
@@ -1,13 +1,16 @@
-import { RegisterU2f } from "./components/register-u2f";
-import { UserAvatar } from "@serverComponents/UserAvatar";
-import { AuthPanel } from "@serverComponents/globals/AuthPanel";
+import { Metadata } from "next";
+import { headers } from "next/headers";
+
 import { getServiceUrlFromHeaders } from "@lib/service-url";
 import { loadSessionById } from "@lib/session";
 import { getSessionCredentials } from "@lib/cookies";
 import { getSerializableLoginSettings } from "@lib/zitadel";
-import { Metadata } from "next";
+
+import { RegisterU2f } from "@components/mfa/u2f/ RegisterU2f";
+import { UserAvatar } from "@serverComponents/UserAvatar";
+import { AuthPanel } from "@serverComponents/globals/AuthPanel";
+
 import { serverTranslation } from "@i18n/server";
-import { headers } from "next/headers";
 
 export async function generateMetadata(): Promise<Metadata> {
   const { t } = await serverTranslation("u2f");
