+++ tsconfig.json
@@ -22,6 +22,7 @@
     "paths": {
       "@lib/*": ["lib/*"],
       "@api/*": ["app/api/*"],
+      "@components/*": ["components/*"],
       "@serverComponents/*": ["components/serverComponents/*"],
       "@clientComponents/*": ["components/clientComponents/*"],
       "@i18n": ["i18n"],
